VFS:uidshift: add flags and helpers to shift UIDs and GIDs to virtual view
Add helpers to allow the VFS to shift UIDs and GIDs according to the
user namespace of the containing mount namespace. All decisions are
taken by VFS.

In this regard, to allow the shift of UID and GID filesystems when
mounted must set "vfs_shift_uids" and "vfs_shift_gids" options,
therefore add the appropriate flags to distinguish between mounts that
support UID/GID shifts and others. The flags are "SB_I_VFS_SHIFT_UIDS"
and "SB_I_VFS_SHIFT_GIDS" and should be set my filesystems in the
super_block->s_iflags when they are mounted.

If a mount supports VFS UID/GID shifts shows up in a mount namespace
that allows UID/GID shifts, VFS helpers will handle the shift
and translate it according to that mount namespace. The user namespace
of that containing mount namespace will be used to perform the
translation, this gives the advantage that only a process with the right
privileges or an ancestor of the mount namespace is able to set user
namespace mappings.

As an example do the shift to the virtual view when we stat() inodes, the
VFS handles that and inodes->{i_uid|i_gid} will always contain the
on-disk view.

Returned stat()->{uid|gid} example inside user_ns_X
---------------------------------------------------

-------------------------------------------------------------------------
inode->uid on Disk | init_user_ns uid | userns_ns_X uid | stat->uid
-------------------------------------------------------------------------
0                  | 1000000          | 0               | 65534
-------------------------------------------------------------------------
999                | 1000999          | 999             | 65534
-------------------------------------------------------------------------
1000               | 1001000          | 1000            | 65534
-------------------------------------------------------------------------
1000000            | 1000000          | 0               | 0
-------------------------------------------------------------------------
1000999            | 1000000          | 0		| 999
-------------------------------------------------------------------------
1001000            | 1000999          | 999		| 1000
-------------------------------------------------------------------------

With this patch:
-------------------------------------------------------------------------
inode->uid on Disk | init_user_ns uid | userns_ns_X uid | stat->uid
-------------------------------------------------------------------------
0                  | 1000000          | 0               | 0
-------------------------------------------------------------------------
999                | 1000999          | 999             | 999
-------------------------------------------------------------------------
1000               | 1001000          | 1000            | 1000
-------------------------------------------------------------------------
1000000            | 1000000          | 0               | 0
-------------------------------------------------------------------------
1000999            | 1000000          | 0		| 999
-------------------------------------------------------------------------
1001000            | 1000999          | 999		| 1000
-------------------------------------------------------------------------

Signed-off-by: Dongsu Park <dongsu@endocode.com>
Signed-off-by: Djalal Harouni <tixxdz@opendz.org>