fs: Treat foreign mounts as nosuid
If a process gets access to a mount from a different mount namespace,
that process should not be able to take advantage of setuid files or
selinux entrypoints from that filesystem.  Prevent this by treating
mounts from other mount namespaces and those not owned by
current_user_ns() or an ancestor as nosuid. This handles the case of
mounts which may be accessed through /proc or if the caller has a
handle on the mount namespace. To be safe a process must clearly
create the appropriate mount namespace and mount or bind mount
mountpoints inside this new mount namespace.

This patch was just adapted from the original one that was written
by Andy Lutomirski <luto@amacapital.net>
https://www.redhat.com/archives/dm-devel/2016-April/msg00374.html

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>