security:task: do not abuse the file system /proc/pid/ API
Currently, processes are able to bypass seccomp filters that block
ptrace and maybe some other direct system calls that operate on
processes/pids by using the indirect way of filesystem calls
open/read/write/close on the /proc/pid/ provided API. LSMs also do
not have a way to restrict this.

There is the hidepid= mount option of /proc which is exclusively a
per-pidns setting, this means that to be able to sandbox some processes
we have also to create a private PID namespace with all the complexity
that it brings.

This patch adds a per-task flag NO_FS_PID_ACCESS that makes any access
using filesystem calls on /proc/pid go through the same permission
checks as ptrace. This reduces the ability of processes to examine or
operate on other processes. The flag can be set without requiring any
privilges.

The per-task flag is an opt-in operation:
        prctl(PR_SET_NO_FS_PID_ACCESS, 1);

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>