proc: allow to change proc mount options per mount
Currently hidepid mount option is propagated to all proc mounts that are
in the same pid namespace. This patch make it possible to have proc
mounts with different options inside the same pid namespace.

Since this may break userspace or code that checks/expects some device
identifier, this mode is only supported when theo ption "version=2" is
passed.

[Buggy patch]

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>